show a test by ID

+++ app/http/controllers/test.controller.js
@@ -18,25 +18,28 @@ class TestContrller {
         }
     }
 
-    async findTest(testID, owner){
-        const test= await TestModel.findOne({owner, _id: testID});
-        if(!test) throw{status: 404, message: "Test does not exist."};
-        
-        return test;
-    }
-
     async editTestById(req, res, next){
         try {
             const owner= req.user._id;
             const testID= req.params.id;
-            const test= await TestModel.findOne({owner})
+            const test= await TestModel.findOne({owner, _id: testID})
             if(!test) throw {status: 404, message: "You cannot access this test to edit."}
-            return res.send("OKAY")
+            // return res.send("OKAY")
             const attendees= test.attendees;
             if(!attendees.length == 0) throw{status:201, message: "Cannot edit a test after it's given."}
             const data= {...req.body};
-            // console.log("owner: ",owner);
-            // console.log("info: ",test);
+
+            Object.entries(data).forEach(([key, value])=>{
+                if(!["title", "introduction", "public"].includes(key)) delete data[key];
+                if(["", " ", 0, null, undefined, NaN].includes(value)) delete data[key];
+            })
+            const updateResult= await TestModel.updateOne({_id: testID}, {$set: data});
+            if(updateResult.modifiedCount == 0) throw {status: 400, message: "Update failed. Please try again."};
+            return res.status(200).json({
+                status: 200,
+                success: true,
+                message:"The test updated successfuly."
+            })
             
         } catch (error) {
             next(error)
@@ -47,7 +50,8 @@ class TestContrller {
         try {
             const owner= req.user._id;
             const testID= req.params.id;
-            await this.findTest(testID, owner);
+            const test= await TestModel.findOne({owner, _id: testID})
+            if(!test) throw {status: 404, message: "You cannot access this test to edit."}            
             const deleteTestResult= await TestModel.deleteOne({_id: testID});
             if(deleteTestResult.deletedCount == 0) throw {status: 400, message: "Failed deleting the test. Plase try again."}
             return res.status(200).json({
@@ -55,6 +59,24 @@ class TestContrller {
                 success: true,
                 message: "The test deleted successfuly."
             })
+        } catch (error) {
+            next(error)
+        }
+    }
+
+    async showATestById(req, res, next){
+        try {
+            const ownerID= req.user._id;
+            const testID= req.params.id;
+            const test= await TestModel.findById(testID);
+            if(!test) throw{status: 404, message: "Test not found."}
+            // console.log(""+ownerID);
+            // console.log(""+test.owner);
+            if(test.public== false){
+                if(""+test.owner== ""+ownerID) return res.json(test)
+
+            } throw{status: 401, message: "This is a private test."}
+            
         } catch (error) {
             next(error)
         }
